requirements.txt
# MIRROR Architecture dependencies
requests>=2.28.0
python-dotenv>=0.20.0 

# CuRaTe dependencies
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.4.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
et-xmlfile==1.1.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.5
idna==3.7
jiter==0.5.0
keyboard==0.13.5
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.0
openai==1.35.10
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
replicate==0.28.0
requests==2.32.3
rsa==4.9
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wordcloud==1.9.3
